requirements-dev.txt
# Tests
dill
filelock
pyrefly
numpy
pre-commit
pytest
pytest-cov
pytest-order
pytest-timeout
pytest-xdist
setuptools
fairlearn
# Test contrib dependencies
scipy
pytorch_fid
tqdm
scikit-learn
matplotlib
tensorboardX
# visdom  # Removed: unmaintained package, cannot be installed with modern packages
polyaxon
wandb
mlflow
neptune-client>=0.16.17
tensorboard
torchvision
pynvml<12  # pynvml module was removed in 12.X, is not developed or maintained. We should replace pynvml with something else.
clearml
scikit-image
pycocotools
# temporary fix for python=3.12 and v3.8.1
# nltk
git+https://github.com/nltk/nltk@aba99c8
# Examples dependencies
pandas
gymnasium
# temporary fix: E   AttributeError: module 'mpmath' has no attribute 'rational'
mpmath<1.4
